Improve seed URL tracking. Changed data type in DB to Int instead of Long because input was Int, too. Also explicitly disallowed duplicates and only decrease the seed counter if removeURL has actually removed a URL from the database to make sure that it does not reach 0 to soon.

+++ src/main/java/edu/uci/ics/crawler4j/frontier/Frontier.java
@@ -164,7 +164,7 @@ public class Frontier extends Configurable {
     return workQueues.getLength() + inProcessPages.getLength();
   }
   
-  public long numOffspring(Integer seedDocid) {
+  public int numOffspring(Integer seedDocid) {
     return workQueues.getSeedCount(seedDocid) + inProcessPages.getSeedCount(seedDocid);
   }
